pyproject.toml
[project]
name = "talent-flow-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Jeferson Ferreira da Silva",email = "jeferson.ferreira@unifesspa.edu.br"}
]
readme = "README.md"
requires-python = ">=3.11.0rc1"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-cov (>=4.1.0,<5.0.0)",
    "numpy (>=1.25.2,<2.6.0) ; python_version >= '3.11.0rc1'",
    "scipy (>=1.8.0,<2.0.0) ; python_version >= '3.11.0rc1'",
    "scikit-learn (>=1.3.0,<2.0.0) ; python_version >= '3.11.0rc1'",
    "pandas (>=2.0.0,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "loguru (>=0.7.0,<0.8.0)",
    "httpx (>=0.27.0,<0.28.0)",
    "requests (>=2.32.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
